refactor 2-sortable-table-v3
