+++ 07-async-code-fetch-api-part-1/2-sortable-table-v3/index.js
@@ -1,11 +1,10 @@
 import fetchJson from './utils/fetch-json.js'
 import createDomElement from './utils/createDomElement.js'
 import MemoDOM from './MemoDom.js'
-import TableTemplate from './TableTemplate.js'
 
 const BACKEND_URL = 'https://course-js.javascript.ru'
 
-export default class SortableTable extends TableTemplate {
+export default class SortableTable {
   #elementDOM = null
   #apiUrl = ''
   #memo = new MemoDOM()
@@ -15,83 +14,20 @@ export default class SortableTable extends TableTemplate {
   data = []
   headerConfig = []
   sorted = {}
+  isSortLocally = false
 
-  constructor(headerConfig = [], 
-    { 
-      url = '', 
-      sorted = {
-        id: headerConfig.find(item => item.sortable).id,
-        order: 'asc'
-      }, 
-      isSortLocally = false 
-    } = {}
-  ) {
-    super()
+  onWindowScroll = async () => {
+    const { bottom } = this.#elementDOM.getBoundingClientRect()
+    const { clientHeight } = document.documentElement
 
-    this.#apiUrl = `${BACKEND_URL}/${url}`
+    if (this.isLoading) return
+    if (bottom >= clientHeight) return
+    if (this.isSortLocally) return
 
-    this.data = []
-    this.sorted = { fieldValue: sorted.id, orderValue: sorted.order }
-    this.headerConfig = headerConfig
-    this.isSortLocally = isSortLocally
-
-    this.loadData({ ...sorted })
-    this.render()
-
+    this.loadMoreData()
   }
 
-  get element() {
-    return this.#elementDOM
-  }
-
-  get subElements() {
-    return this.#memo.cache
-  }
-
-  render() {
-    this.#elementDOM = createDomElement(this.buildTemplate())
-    this.#memo.memoizeDocument(this.#elementDOM)
-    this.initColumnsActions()
-    this.initScrollAction()
-  }
-
-  remove() {
-    this.#elementDOM?.remove()
-  }
-
-  destroy() {
-    this.remove()
-    this.#elementDOM = null
-    this.headerConfig = null
-    this.data = null
-
-    this.#memo.clear()
-    document.removeEventListener('scroll', this.handleScroll)
-  }
-
-  initScrollAction() {
-    document.addEventListener('scroll', this.handleScroll)
-  }
-
-  initColumnsActions() {
-    const cacheDOM = this.#memo.cache
-    cacheDOM.header.addEventListener('click', this.handleSort)
-  }
-
-  handleScroll = async (event) => {
-    if (this.isLoading) return 
-    const block = this.#memo.cache.body
-    const contentHeight = block.offsetHeight
-    const yOffset = window.pageYOffset
-    const window_height = window.innerHeight
-    const y = yOffset + window_height
-    
-    if(!(y >= contentHeight)) return
-
-    this.loadMoreData() 
-  }
-
-  handleSort = (event) => {
+  onSortClick = (event) => {
     const col = event.target.closest('[data-sortable=true]')
     if (!col) return 
 
@@ -113,6 +49,69 @@ export default class SortableTable extends TableTemplate {
     this.loadData()
   }
 
+  constructor(headerConfig = [], 
+    { 
+      url = '', 
+      sorted = {
+        id: headerConfig.find(item => item.sortable).id,
+        order: 'asc'
+      }, 
+      isSortLocally = false 
+    } = {}
+  ) {
+    this.#apiUrl = `${BACKEND_URL}/${url}`
+
+    this.data = []
+    this.sorted = { fieldValue: sorted.id, orderValue: sorted.order }
+    this.headerConfig = headerConfig
+    this.isSortLocally = isSortLocally
+
+    this.loadData({ ...sorted })
+    this.render()
+  }
+
+  get element() {
+    return this.#elementDOM
+  }
+
+  get subElements() {
+    return this.#memo.cache
+  }
+
+  render() {
+    this.#elementDOM = createDomElement(this.buildTemplate())
+    this.#memo.memoizeDocument(this.#elementDOM)
+    this.initEventListeners()
+  }
+
+  remove() {
+    this.#elementDOM?.remove()
+  }
+
+  destroy() {
+    this.remove()
+    this.#elementDOM = null
+    this.headerConfig = null
+    this.data = null
+
+    this.#memo.clear()
+    document.removeEventListener('scroll', this.onSortClick)
+    window.removeEventListener('scroll', this.onWindowScroll)
+  }
+
+  initEventListeners() {
+    this.#memo.cache.header.addEventListener('click', this.onSortClick)
+    window.addEventListener('scroll', this.onWindowScroll)
+  }
+
+  async loadData() {
+    const { fieldValue, orderValue } = this.sorted
+    const request = `${this.#apiUrl}?_sort=${fieldValue}&_order=${orderValue}&_start=0&_end=30`
+    const data = await fetchJson(request)
+    this.data = data
+    this.updateTableBody()
+  }
+
   async loadMoreData() {
     this.isLoading = true
     const end = this.data.length + 30
@@ -131,14 +130,6 @@ export default class SortableTable extends TableTemplate {
     this.#memo.cache.body.append(...newRowsWrapper.children)
 
     this.isLoading = false
-  }
-
-  async loadData() {
-    const { fieldValue, orderValue } = this.sorted
-    const request = `${this.#apiUrl}?_sort=${fieldValue}&_order=${orderValue}&_start=0&_end=30`
-    const data = await fetchJson(request)
-    this.data = data
-    this.updateTableBody()
   }
 
   updateTableBody() {
@@ -171,5 +162,63 @@ export default class SortableTable extends TableTemplate {
       if (elemKey.startsWith('sort-cell'))
         elemDOM.dataset.order = ''
     })
+  }
+
+  buildTemplate() {
+    return /*html*/`
+      <div class="sortable-table">
+
+        <div data-memo="header" class="sortable-table__header sortable-table__row">
+          ${this.buildTemplateHeader()}
+        </div>
+        
+        <div data-memo="body" class="sortable-table__body">
+          ${this.buildTemplateBody()}
+        </div>
+      </div>
+    `
+  }
+
+  buildTemplateBody(customData) {
+    const buildData = customData || this.data
+    return buildData.map((product) => this.buildTemplateRow(product)).join('')
+  }
+
+  buildTemplateHeader() {
+    return this.headerConfig.map(({ id, title, sortable } = {}) => (
+      /*html*/`
+      <div 
+        class="sortable-table__cell"
+        data-memo="sort-cell-${id}"
+        data-id="${id}" 
+        data-sortable="${sortable}" 
+        data-order="${
+          this.sorted.fieldValue === id 
+            ? this.sorted.orderValue 
+            : ''
+        }"
+      >
+        <span>${title}</span>
+        <span data-element="arrow" class="sortable-table__sort-arrow">
+          <span class="sort-arrow"></span>
+        </span>
+      </div>`
+    )).join('') 
+  }
+
+  buildTemplateRow(product) {
+    const row = this.headerConfig.map((col) => {
+      if (!product[col.id]) return
+      if (col.template) return col.template(product[col.id])
+      return /*html*/`
+        <div class="sortable-table__cell">${product[col.id]}</div>
+      `
+    }).join("")
+
+    return /*html*/`
+      <a href="/products/${product.id}" class="sortable-table__row">
+        ${row}
+      </a>
+    `
   }
 }
